pyproject.toml
[project]
name = "task-flow-api-2"
version = "0.1.0"
description = "Async FastAPI eCommerce backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "celery>=5.4.0",
    "email-validator>=2.2.0",
    "fastapi>=0.116.0",
    "greenlet>=3.1.1",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pydantic-settings>=2.8.0",
    "pyjwt>=2.10.1",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "psycopg2-binary>=2.9.10",
    "python-multipart>=0.0.20",
    "redis>=5.2.1",
    "sqlalchemy>=2.0.36",
    "slowapi>=0.1.9",
    "uvicorn[standard]>=0.34.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
